# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: SubnetIQ REST and Streaming API
  version: 1.0.0
  description:
    'Guest mathematics and lookups; Supabase bearer-token persistence; typed assistant SSE. Requests
    accept application/json and are limited to 512 KiB, except project POST/PATCH envelopes (2.25 MiB) with
    a 2 MiB plan cap. Success bodies are returned directly. Failure bodies follow Error. API responses default
    to Cache-Control: no-store. Global quota is 180 requests/minute per IP identity. Production quotas are
    atomic PostgreSQL RPC calls. No credentials means honest guest mode in development; protected operations
    return 503. All ordinary user database access retains the verified user RLS context.'
servers:
  - url: http://localhost:3001
    description: Local API
  - url: /
    description: Current deployment
...
paths:
...
  /api/projects:
    get:
      operationId: get_projects
      summary: List own projects
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - &id002
          name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 200
        - &id003
          name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 100000
    post:
      operationId: post_projects
      summary: Create a project
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
      description: Project request envelope is limited to 2.25 MiB; the plan JSON is limited to 2 MiB.
  /api/projects/{id}:
    get:
      operationId: get_projects_id
      summary: Read own project
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: &id001
            type: string
            format: uuid
    patch:
      operationId: patch_projects_id
      summary: Atomically update and version a project
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Stale version; reload before saving.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        version is the expected current version. The database captures the previous snapshot and
        increments version in the same transaction. The project request envelope is limited to 2.25 MiB and
        plan JSON to 2 MiB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProject'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
    delete:
      operationId: delete_projects_id
      summary: Delete own project and its children
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '204':
          description: Deleted or revoked; response body is empty.
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/projects/{id}/revisions:
    get:
      operationId: get_projects_id_revisions
      summary: List project revisions
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Revision'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
        - *id002
        - *id003
    post:
      operationId: post_projects_id_revisions
      summary: Capture a manual checkpoint
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  default: Manual checkpoint
              additionalProperties: false
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/projects/{id}/restore:
    post:
      operationId: post_projects_id_restore
      summary: Restore a saved revision atomically
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Stale version; reload before restoring.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Send version to detect concurrent editing. If omitted, the API reads the current version
        and still uses an atomic compare-and-update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                revisionId: *id001
                version:
                  type: integer
                  minimum: 1
              required:
                - revisionId
              additionalProperties: false
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/projects/{id}/shares:
    get:
      operationId: get_projects_id_shares
      summary: List share metadata without tokens
      tags:
        - Sharing
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Share'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
    post:
      operationId: post_projects_id_shares
      summary: Publish an expiring immutable snapshot
      tags:
        - Sharing
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewShare'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        A cryptographically random token is returned once. Only its SHA-256 hash is stored; the
        current project is captured by the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInDays:
                  type: integer
                  minimum: 1
                  maximum: 90
                  default: 30
              additionalProperties: false
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/shares/{id}:
    delete:
      operationId: delete_shares_id
      summary: Revoke a share
      tags:
        - Sharing
      security:
        - supabaseBearer: []
      responses:
        '204':
          description: Deleted or revoked; response body is empty.
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/share/{token}:
    get:
      operationId: get_share_token
      summary: Resolve a public snapshot
      tags:
        - Sharing
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invalid, expired, revoked, or missing link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Expires and revocation are checked in the service-only RPC. Responses use no-store and no-referrer.
        The same 404 is returned for invalid, expired, and revoked tokens.
      parameters:
        - name: token
          in: path
          required: true
          schema: &id004
            type: string
...
  /api/saved-calculations/{id}:
    delete:
      operationId: delete_saved_calculations_id
      summary: Delete a saved calculation
      tags:
        - Calculations
      security:
        - supabaseBearer: []
      responses:
        '204':
          description: Deleted or revoked; response body is empty.
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
...
components:
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Supabase access JWT
      description:
        A verified Supabase access token; API calls auth.getUser to verify it before creating a
        user-scoped database client.
  schemas:
...
    Project:
      type: object
      properties:
        id: *id001
        name: &id005
          type: string
          minLength: 1
          maxLength: 100
        description: &id006
          type: string
          maxLength: 3000
        address_space: &id007
          type: string
          maxLength: 100
        plan: &id008
          type: object
          additionalProperties: true
        version: &id010
          type: integer
          minimum: 1
        archived: &id009
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - address_space
        - plan
        - version
        - created_at
        - updated_at
      additionalProperties: true
...